+++ src/main/java/org/openwebnet4j/communication/USBConnector.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2023 Contributors to the openwebnet4j project
+ * Copyright (c) 2020-2024 Contributors to the openwebnet4j project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -15,9 +15,15 @@
 package org.openwebnet4j.communication;
 
 import java.io.IOException;
-import java.util.Enumeration;
-import java.util.TooManyListenersException;
+import java.util.stream.Collectors;
 
+import org.eclipse.jdt.annotation.NonNull;
+import org.eclipse.jdt.annotation.Nullable;
+import org.openwebnet4j.communication.serial.SerialPortProviderSPIFactory;
+import org.openwebnet4j.communication.serial.spi.SerialPort;
+import org.openwebnet4j.communication.serial.spi.SerialPortEvent;
+import org.openwebnet4j.communication.serial.spi.SerialPortEventListener;
+import org.openwebnet4j.communication.serial.spi.SerialPortProvider;
 import org.openwebnet4j.message.AckOpenMessage;
 import org.openwebnet4j.message.Automation;
 import org.openwebnet4j.message.BaseOpenMessage;
@@ -30,15 +36,6 @@ import org.openwebnet4j.message.UnsupportedFrameException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import gnu.io.CommPortIdentifier;
-import gnu.io.NRSerialPort;
-import gnu.io.NoSuchPortException;
-import gnu.io.PortInUseException;
-import gnu.io.SerialPort;
-import gnu.io.SerialPortEvent;
-import gnu.io.SerialPortEventListener;
-import gnu.io.UnsupportedCommOperationException;
-
 /**
  * Class for communicating with a ZigBee USB Gateway using the OpenWebNet protocol
  *
@@ -47,6 +44,7 @@ import gnu.io.UnsupportedCommOperationException;
 public class USBConnector extends OpenConnector implements SerialPortEventListener {
 
     private static final int SERIAL_SPEED = 19200; // UART baud as declared in the OWN specs
+    private static final int SERIAL_PORT_OPEN_TIMEOUT = 1000;
 
     private final Logger logger = LoggerFactory.getLogger(USBConnector.class);
     private final Logger msgLogger = LoggerFactory.getLogger(logger.getName() + ".message");
@@ -62,8 +60,10 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
     private static final String OLD_FIRMWARE_VERSION = "1.2.3"; // firmware versions <= than this are affected by
     // Dimension response bug
 
-    private final String portName; // the serial port name we are connecting to
-    private NRSerialPort serialPort; // the serial port, once connected
+    private final String portName;
+    private SerialPort serialPort;
+
+    private SerialPortProvider serialPortProvider;
 
     private Response currentResponse;
     private final Object requestSentSynchObj = new Object(); // Synch object to synchronise sending a request frame and
@@ -72,6 +72,24 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
     public USBConnector(String portName) {
         super();
         this.portName = portName;
+    }
+
+    /**
+     * Sets a {@link SerialPortProvider} to use to obtain available serial ports
+     *
+     * @param provider the {@link SerialPortProvider} to set
+     */
+    public void setSerialPortProvider(@NonNull SerialPortProvider provider) {
+        serialPortProvider = provider;
+    }
+
+    /**
+     * Gets the current {@link SerialPortProvider}
+     *
+     * @return {@link SerialPortProvider} currently set
+     */
+    public @Nullable SerialPortProvider getSerialPortProvider() {
+        return serialPortProvider;
     }
 
     @Override
@@ -111,6 +129,7 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
 
     private void checkFirmwareVersion() throws OWNException {
         try {
+
             Response res = sendCommandSynchInternal(GatewayMgmt.requestFirmwareVersion().getFrameValue());
             if (res != null) {
                 OpenMessage msg = res.getResponseMessages().get(0);
@@ -153,12 +172,18 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
     private void connectUSBDongle(String portN) throws OWNException {
         if (serialPort == null) {
             serialPort = connectSerialPort(portN);
-            cmdChannel = new FrameChannel(serialPort.getInputStream(), serialPort.getOutputStream(), "USB");
+            try {
+                cmdChannel = new FrameChannel(serialPort.getInputStream(), serialPort.getOutputStream(), "USB");
+            } catch (IOException e) {
+                disconnectSerialPort();
+                throw new OWNException(
+                        "Failed to getInput/OutputStreams for Zigbee USB Gateway on serial port: " + portN, e);
+            }
         }
         try {
-            // send requestKeepConnect to see if USB stick is ready to receive commands
+            // send requestKeepConnect (*13*60*##) to see if USB stick is ready to receive commands
             GatewayMgmt frame = GatewayMgmt.requestKeepConnect();
-            cmdChannel.sendFrame(GatewayMgmt.requestKeepConnect().getFrameValue());
+            cmdChannel.sendFrame(frame.getFrameValue());
 
             hsLogger.info("(HS) USB HS==>>>> `{}`", frame.getFrameValue());
             Thread.sleep(50); // we must wait few ms for the answer to be ready
@@ -170,123 +195,75 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
                         + ". Serial returned: " + resp);
             }
             // set event listener for incoming frames
-            serialPort.addEventListener(this);
-            logger.debug("##USB-conn## added event listener");
-            logger.info("##USB-conn## === CONNECTED TO USB GATEWAY on serial port: {} ===", portN);
+            if (serialPort.addEventListener(this)) {
+                logger.debug("##USB-conn## added event listener");
+                logger.info("##USB-conn## === CONNECTED TO USB GATEWAY on serial port: {} ===", portN);
+            } else {
+                disconnectSerialPort();
+                throw new OWNException("Failed to communicate with Zigbee USB Gateway on serial port " + portN
+                        + ": could not add event listener");
+            }
         } catch (InterruptedException e) {
             e.printStackTrace();
         } catch (IOException e) {
             disconnectSerialPort();
             throw new OWNException("Failed to communicate with Zigbee USB Gateway on serial port: " + portN, e);
-        } catch (TooManyListenersException e) {
-            disconnectSerialPort();
-            throw new OWNException("Failed to communicate with Zigbee USB Gateway on serial port: " + portN, e);
         }
     }
 
-    private NRSerialPort connectSerialPort(String portN) throws OWNException {
-        try {
-            // FIXME If I call RXTXVersion.getVersion(), it throws NoClassDefError each time gnu.io
-            // is called again
-            // logger.debug("RXTXVersion: {}", RXTXVersion.getVersion());
-        } catch (NoClassDefFoundError e) {
-            logger.error(
-                    "##USB-conn## Serial connection requires RXTX libraries to be available, but they could not be found!");
-            throw new OWNException(
-                    "Serial connection requires RXTX libraries to be available, but they could not be found!", e);
-        }
-        CommPortIdentifier ident = null;
-        try { // see if serial port exists
-            ident = CommPortIdentifier.getPortIdentifier(portN);
-        } catch (NoSuchPortException e) {
-            logger.info("##USB-conn## Failed to connect to serial port {}: NoSuchPortException", portN);
-            String availPorts = listSerialPorts();
-            logger.info("##USB-conn## Available serial ports are: {}", availPorts);
-            throw new OWNException(
-                    "Failed to connect to serial port " + portN + ". Available serial ports are: " + availPorts, e);
-        }
-        logger.debug("##USB-conn## CommPortIndetifier: name={} type={} owner={}", ident.getName(), ident.getPortType(),
-                ident.getCurrentOwner());
+    private SerialPort connectSerialPort(String portN) throws OWNException {
+        if (serialPortProvider == null) {
+            logger.debug("##USB-conn## No SerialPortProvider set, getting it via SPI/default...");
 
-        if (ident.getPortType() != CommPortIdentifier.PORT_SERIAL) {
-            logger.error("##USB-conn## Port {} is not a serial port", ident.getName());
-            throw new OWNException("Failed to connect to port " + portN + " (not a serial port).");
+            SerialPortProviderSPIFactory portManager = new SerialPortProviderSPIFactory();
+            serialPortProvider = portManager.getProvider();
         }
-        if (ident.isCurrentlyOwned()) {
-            logger.debug("##USB-conn## Serial port {} is already in use", ident.getName());
-            throw new OWNException("Failed to connect to serial port " + portN + ". Port is already in use.");
-        }
-        NRSerialPort connectPort = new NRSerialPort(portN, SERIAL_SPEED);
-        logger.debug("##USB-conn## NRSerialPort created");
 
-        try {
-            if (!connectPort.connect()) {
-                logger.info("Failed to connect to serial port {}", ident.getName());
-                throw new OWNException("Failed to connect to serial port " + ident.getName()
-                        + " (NRSerialPort.connect() returned false)");
-            }
-            connectPort.getSerialPortInstance().setSerialPortParams(SERIAL_SPEED, SerialPort.DATABITS_8,
-                    SerialPort.STOPBITS_1, SerialPort.PARITY_NONE);
-        } catch (UnsupportedCommOperationException ue) {
-            logger.error("##USB-conn## Failed setting params for port {}", ident.getName());
-            throw new OWNException("Failed setting params for port: " + portN, ue);
-        } catch (OWNException oe) {
-            throw oe;
-        } catch (Exception e) {
-            // PortInUseException could be thrown but is not declared, so we check here
-            if (e instanceof PortInUseException) {
-                logger.debug("##USB-conn## Serial port {} is already in use (PortInUseException)", ident.getName());
-                throw new OWNException(
-                        "Failed to connect to serial port " + portN + ". Port is already in use (PortInUseException).",
-                        e);
+        SerialPort tempSp = null;
+        // see if serial port exists
+        tempSp = serialPortProvider.getSerialPort(portN);
+        if (tempSp == null) {
+            logger.warn("##USB-conn## Failed to connect to serial port {}: port cannot be found", portN);
+            String availPorts = listSerialPorts(serialPortProvider);
+            logger.warn("##USB-conn## Available serial ports are: {}", availPorts);
+            throw new OWNException(
+                    "Failed to connect to serial port " + portN + ". Available serial ports are: " + availPorts);
+        } else {
+            if (tempSp.open()) {
+                logger.debug("##USB-conn## [{}] SerialPort connected", portN);
+                if (tempSp.setSerialPortParams(SERIAL_SPEED, SerialPort.DATABITS_8, SerialPort.STOPBITS_1,
+                        SerialPort.PARITY_NONE)) {
+                    logger.debug("##USB-conn## [{}] SerialPort connection params set", portN);
+                    logger.info("##USB-conn## [{}] === CONNECTED TO SERIAL PORT ===", tempSp.getName());
+                    return tempSp;
+                } else {
+                    logger.error("##USB-conn## [{}] Failed setting params for serial port", portN);
+                    throw new OWNException("Failed setting params for port: " + portN);
+                }
             } else {
-                logger.debug("##USB-conn## Exception while connecting serial port: {}", e.getMessage());
-                throw new OWNException("Failed to connect to serial port " + portN, e);
+                logger.error("##USB-conn## [{}] Could not open serial port", portN);
+                throw new OWNException("Failed to open serial port " + portN + " (port already in use?)");
             }
         }
-        SerialPort sp = connectPort.getSerialPortInstance();
 
-        // try {
-        // sp.enableReceiveThreshold(1); // makes read() blocking until at least 1 char is
-        // received
-        // sp.disableReceiveTimeout(); // disable any read() timeout
-        // sp.enableReceiveTimeout(SERIAL_RECEIVE_TIMEOUT);
-        logger.debug("##USB-conn## isReceiveThresholdEnabled={} v={}", sp.isReceiveThresholdEnabled(),
-                sp.getReceiveThreshold());
-        logger.debug("##USB-conn## isReceiveTimeoutEnabled={} v={}", sp.isReceiveTimeoutEnabled(),
-                sp.getReceiveTimeout());
-        logger.debug("##USB-conn## FlowControl: {}", sp.getFlowControlMode());
-
-        logger.info("##USB-conn## === CONNECTED TO SERIAL PORT {} ===", connectPort.getSerialPortInstance().getName());
-        return connectPort;
     }
 
     private void disconnectSerialPort() {
         if (serialPort != null) {
-            serialPort.disconnect();
+            serialPort.close();
             logger.debug("##USB-conn## Serial port {} DISCONNECTED", portName);
             serialPort = null;
         }
     }
 
-    private String listSerialPorts() {
-        StringBuilder sb = new StringBuilder();
-        Enumeration<?> portList = CommPortIdentifier.getPortIdentifiers();
-        boolean first = true;
-        while (portList.hasMoreElements()) {
-            CommPortIdentifier id = (CommPortIdentifier) portList.nextElement();
-            if (id.getPortType() == CommPortIdentifier.PORT_SERIAL) {
-                if (!first) {
-                    sb.append(", ");
-                }
-                sb.append(id.getName());
-                first = false;
-            }
+    private String listSerialPorts(@NonNull SerialPortProvider provider) {
+        String result = provider.getSerialPorts().map(sp -> sp.getName()).collect(Collectors.joining(", "));
+        if (result.length() == 0) {
+            return "-NO SERIAL PORTS FOUND-";
+        } else {
+            return result;
         }
-        if (sb.length() == 0) {
-            sb.append("-NO SERIAL PORTS FOUND-");
-        }
-        return sb.toString();
+
     }
 
     @Override
@@ -393,36 +370,36 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
     }
 
     @Override
-    public void serialEvent(SerialPortEvent event) {
+    public void serialEvent(@NonNull SerialPortEvent event) {
         int eventType = event.getEventType();
         switch (eventType) {
-            case SerialPortEvent.DATA_AVAILABLE:
+            case SerialPortEvent.EVENT_DATA_AVAILABLE:
                 String frame = null;
-                logger.debug("##USB-conn## START Processing DATA_AVAILABLE event...");
+                logger.debug("##USB-conn## START Processing EVENT_DATA_AVAILABLE...");
                 do {
                     try {
                         frame = cmdChannel.readFrames();
                     } catch (IOException e) {
-                        logger.error("##USB-conn## IOException while reading frames from DATA_AVAILABLE event: {}",
+                        logger.error("##USB-conn## IOException while reading frames from EVENT_DATA_AVAILABLE: {}",
                                 e.getMessage());
                     }
                     if (frame == null) {
                         logger.trace(
-                                "##USB-conn## no more frames to read from DATA_AVAILABLE event (readFrames() returned {})",
+                                "##USB-conn## no more frames to read from EVENT_DATA_AVAILABLE (readFrames() returned {})",
                                 frame);
                     } else {
                         processFrame(frame);
                     }
                 } while (frame != null);
-                logger.trace("##USB-conn## END processing DATA_AVAILABLE event");
+                logger.trace("##USB-conn## END processing EVENT_DATA_AVAILABLE");
                 break;
-            case 11: // Event is defined from nrjavaserial > 5.1.0 - 11 =
-                // SerialPortEvent.HARDWARE_ERROR
-                logger.warn("##USB-conn## serialEvent received HARDWARE_ERROR event: disconnecting serial port {}...",
+            case SerialPortEvent.EVENT_PORT_DISCONNECTED:
+                logger.warn(
+                        "##USB-conn## serialEvent received EVENT_PORT_DISCONNECTED: disconnecting serial port {}...",
                         portName);
                 disconnectCmdChannel();
                 disconnectSerialPort();
-                handleMonDisconnect(new OWNException("Serial port " + portName + " received HARDWARE_ERROR event"));
+                handleMonDisconnect(new OWNException("Serial port " + portName + " received EVENT_PORT_DISCONNECTED"));
                 break;
             default:
                 logger.debug("##USB-conn## serialEvent() received unhandled event type: {}", event.getEventType());
@@ -467,4 +444,5 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
         // e.g. "1.2.3" = "1.2.3" or "1.2.3" < "1.2.3.4"
         return Integer.signum(vals1.length - vals2.length);
     }
+
 }